hair_transfer/app/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
Pillow>=10.0.0
PyYAML>=6.0.1
